create super user when adding the auth controller
